Feat: Email templates from sendgrid + custom fields
