+++ server/app.js
@@ -7,6 +7,7 @@ var cors = require('cors');
 var mongoose = require("mongoose");
 var app = express();
 var bcrypt = require('bcryptjs');
+var fs = require("fs");
 
 const uri =
   "mongodb+srv://Admin:DataStructures@cluster0-wcree.mongodb.net/TutorBase?retryWrites=true&w=majority";
@@ -19,7 +20,7 @@ var indexRouter = require("./routes/index");
 var usersRouter = require("./routes/users");
 var signupRouter = require("./routes/signup");
 var loginRouter = require("./routes/login");
-var emailClientRouter = require("./routes/email-client");
+var emailClientRouter = require("./routes/email-user");
 
 // app.use(cors({
 //   origin: "http://localhost:3000",
@@ -41,7 +42,7 @@ app.use("/", indexRouter)
 app.use("/users", usersRouter);
 app.use("/login", loginRouter);
 app.use("/signup", signupRouter);
-app.use("/email-client", emailClientRouter);
+app.use("/email-user", emailClientRouter);
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
